usb: gadget: ffs: Fix runtime PM usage count for multiple set_alt
In some scenarios such as USBCV tests, multiple set_alt
could be issued with same interface and alternate
setting. In that case, EPs are disabled and enabled,
we get the usage count but do not put it. This causes
a mismatch and DWC3 LPM does not work anymore. Fix
this issue by moving put_usage_count to ffs_func_set_alt.

Change-Id: Ic85db09820d2c18b67bd463fe27e2c1ef41e8c74
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>